Add a few prototypes for the UBSAN interfaces.
This required adding a new alias Die() to common/, which aborts the program.
In addition, the alias situation was faulty before and would cause compilation
errors when included directly. This has been fixed by moving the alias to the
compilation unit where the implementation is declared. The headers now just
declare the external symbol.