Add __ubsan_handle_dynamic_type_cache_miss functions
This commit stubs the handlers that get called when attempting to retrieve
RTTI for an object that is not in the 'caches' LLVM keeps around at runtime.
In most cases, this reflects corrupted vptrs and may crash ubsan (ironic), but
our stub should be safe because it does absolutely nothing useful.

Fixes #3