Add controlling macros for the logger and die modules
These modules are needed to provide sanitizer functionality, but the APIs they
rely on may not exist on some obscure systems. In order to allow people to
build the library for such systems without having to provide stubs,
configuration macros were added. If these macros are not enabled, minimal
stubs will be compiled instead to prevent linking errors if an application
doesn't override them with its own implementation.

Additionally, a rogue use of the printf() function was mercilessly eliminated.