Rewrite the test infrastructure to use inline tests
This encodes the test parameters into the source code itself so we don't
have to keep multiple files in sync to make tests pass. Additionally,
it makes having multiple tests far less fragile and makes re-using
test cases from compiler-rt a lot more straightforward.

The way it works is that comments of the form
/* CHECK: <regex> */
are parsed and the regex compared against stderr, where sanitizers output their
information by default. If the output is found in stderr, the test passes.
Otherwise, the test fails.